Move profile notify edit component parts to separate files
