+++ client/src/app/pages/dashboard/profile/profile-follow/profile-notify-edit/profile-notify-edit.component.html
@@ -1,0 +1,40 @@
+<nb-card>
+  <nb-card-header>
+    Edit Notification Settings
+  </nb-card-header>
+  <nb-card-body>
+    Notify me when this user<br><br>
+    <ul class="list-unstyled">
+      <li>
+        <nb-checkbox [(ngModel)]="checkboxFlags.PB.checked">
+          achieves a personal best
+        </nb-checkbox>
+      </li>
+      <li>
+        <nb-checkbox [(ngModel)]="checkboxFlags.WR.checked">
+          achieves a world record
+        </nb-checkbox>
+      </li>
+      <li>
+        <nb-checkbox [(ngModel)]="checkboxFlags.MAP_APPROVED.checked">
+          gets a map approved
+        </nb-checkbox>
+      </li>
+      <li>
+        <nb-checkbox [(ngModel)]="checkboxFlags.MAP_UPLOADED.checked">
+          uploads a map
+        </nb-checkbox>
+      </li>
+    </ul>
+  </nb-card-body>
+  <nb-card-footer>
+    <div class="row no-gutters">
+      <div class="col-6 pr-1">
+        <button class="btn btn-danger m-0 w-100" (click)="close()">Close</button>
+      </div>
+      <div class="col-6 pl-1">
+        <button class="btn btn-primary m-0 w-100" (click)="submit()">Submit</button>
+      </div>
+    </div>
+  </nb-card-footer>
+</nb-card>
