+++ client/src/app/pages/dashboard/profile/profile-follow/profile-notify-edit/profile-notify-edit.component.scss
@@ -1,0 +1,6 @@
+button {
+  margin: 1rem;
+}
+nb-card {
+  min-width: 300px;
+}
