+++ client/src/app/pages/dashboard/profile/profile-follow/profile-notify-edit/profile-notify-edit.component.spec.ts
@@ -1,0 +1,31 @@
+// import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+//
+// import { ProfileNotifyEditComponent } from './profile-notify-edit.component';
+// import {NbCardModule, NbCheckboxModule, NbThemeModule} from "@nebular/theme";
+//
+// describe('ProfileNotifyEditComponent', () => {
+//   let component: ProfileNotifyEditComponent;
+//   let fixture: ComponentFixture<ProfileNotifyEditComponent>;
+//
+//   beforeEach(async(() => {
+//     TestBed.configureTestingModule({
+//       imports: [
+//         NbThemeModule,
+//         NbCardModule,
+//         NbCheckboxModule,
+//       ],
+//       declarations: [ ProfileNotifyEditComponent ],
+//     })
+//     .compileComponents();
+//   }));
+//
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(ProfileNotifyEditComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
+//
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
