+++ client/src/app/pages/dashboard/profile/profile.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {UserProfileComponent} from './user-profile.component';
 import {ProfileEditComponent} from './profile-edit/profile-edit.component';
-import {ProfileFollowComponent, ProfileNotifyEditComponent} from './profile-follow/profile-follow.component';
+import {ProfileFollowComponent} from './profile-follow/profile-follow.component';
 import {ProfileCreditsComponent} from './profile-credits/profile-credits.component';
 import {ProfileRunHistoryComponent} from './profile-run-history/profile-run-history.component';
 import {ProfileRoutingModule} from './profile-routing.module';
@@ -14,6 +14,7 @@ import {MarkdownModule} from 'ngx-markdown';
 import {FormsModule} from '@angular/forms';
 import {NgxPaginationModule} from 'ngx-pagination';
 import {UserModule} from '../user/user.module';
+import {ProfileNotifyEditComponent} from './profile-follow/profile-notify-edit/profile-notify-edit.component';
 
 @NgModule({
   imports: [
@@ -33,9 +34,9 @@ import {UserModule} from '../user/user.module';
     UserProfileComponent,
     ProfileEditComponent,
     ProfileFollowComponent,
-    ProfileNotifyEditComponent,
     ProfileCreditsComponent,
     ProfileRunHistoryComponent,
+    ProfileNotifyEditComponent,
   ],
   providers: [],
   entryComponents: [
